+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.3.5
+
+- Estimate peg-in transaction size, total fee and PPM
+- Add peer fee revenue stats to the main page
+
 ## 1.3.4
 
 - Display channel flows since their last swaps